Added the Floor Sequencer. Updated ThetaStar.
Removed searchArea from being a member. Fixed a bug where the start of
findPath in GridAStar and ThetaStar was not resetting the search area
with the appropriate parent, g, and h values and was using the values
from a previous call to findPath. Added a FloorConnectorNode to subclass
the GridNode. The FloorConnectorNode represents staircases and elevators
in the building.